requirements.txt
pyspark==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
loguru==0.7.0
pydantic==2.4.2
pydantic-settings==2.0.3
jupyter==1.0.0
notebook==6.5.4